samples/acme-compose/app/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pytest>=7.0.0
httpx>=0.26.0
